pyproject.toml
[project]
name = "denario"
version = "0.1.7"
description = "Modular Automation of Scientific Research with Multi-Agent Systems"
authors = [
  { name = "Boris Bolliet", email = "boris.bolliet@gmail.com" },
  { name = "Francisco Villaescusa-Navarro", email = "villaescusa.francisco@gmail.com" },
  { name = "Pablo Villanueva-Domingo", email = "pablo.villanueva.domingo@gmail.com" }, 
]
maintainers = [
  { name = "Pablo Villanueva-Domingo", email = "pablo.villanueva.domingo@gmail.com" },
  { name = "Francisco Villaescusa-Navarro", email = "villaescusa.francisco@gmail.com" },
  { name = "Boris Bolliet", email = "boris.bolliet@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
license = { text = "GPLv3" }
dependencies = [
  "langchain>=0.3.23",
  "langchain-chroma>=0.2.2",
  "langchain-community>=0.3.21",
  "langchain-google-vertexai>=2.0.19",
  "langchain-groq>=0.3.2",
  "langchain-google-genai",
  "langchain-anthropic>=0.3.10",
  "langchain-openai>=0.3.12",
  "openai==1.99.9",
  "python-dotenv>=1.0.1",
  "langgraph>=0.3.25",
  "google-cloud-aiplatform",
  "google-ai-generativelanguage>=0.6.17",
  "flatbuffers>=24.12.23",
  "filelock>=3.17.0",
  "cmbagent>=0.0.1post53",
  "pymupdf",
  "pillow",
  "jsonschema",
  "futurehouse-client"
]

[project.urls]
homepage = "https://astropilot-ai.github.io/DenarioPaperPage/"
documentation = "https://denario.readthedocs.io/en/latest/"

[dependency-groups]
dev = [
  {include-group = "docs"},
  {include-group = "app"}
]
docs = [
  "mkdocs",
  "mkdocstrings[python]",
  "mkdocs-material>=9.6.12",
  "mkdocs-jupyter>=0.25.1",
  "mkdocs-git-committers-plugin-2>=2.5.0",
]
app = [
  "streamlit",
  "streamlit_pdf_viewer",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.setuptools.packages.find]
where = ["."]
